IDB: Implement IndexedDB database clean up during an active session.
The previous database cleaner (tombstone sweeper and db compaction)
used to occur only during database connection closure. This caused a
variety of performance issue where the tombstones could pile up within
the session or would not be cleaned as the browser would also close
when the connection closed.

This CL introduces a clean up routine which can run while the database
connection is active. The routine consists of both tombstone sweeping
and compaction, and will be triggered (after a delay) any time a cursor
encounters enough tombstones while iterating.

The task for clean up will be postponed if there are any active
transactions. We will also trigger the run regardless of oncoming calls
when an upper limit of the time has passed since the last run.

Bug: 374691835
Change-Id: I4596514086ccb1e045f573d25095c3f3cd273de2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5948643
Reviewed-by: Abhishek Shanthkumar <abhishek.shanthkumar@microsoft.com>
Reviewed-by: Steve Becker <stevebe@microsoft.com>
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Commit-Queue: Tanuj Martolia <tamartol@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1418535}