Simplify compile_multilib logic
Instead of the awkward and brittle "only output compile_multilib if
there are other fields" logic, explicitly set the multilib attribute
where it's needed.

This has the effect of removing the compile_multilib attribute for all
modules except cc_defaults. Which seems to be enough for it to keep
working.

Also, document why we are doing this because the failure this remedies
is quite obscure.

Bug: b:394038382
Change-Id: I6c571e52c6549227035cbe2dc837e0966127ded4
Cq-Include-Trybots: luci.chromium.try:android-cronet-arm64-gn2bp-dbg
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6249858
Reviewed-by: Mohannad Farrag <aymanm@google.com>
Commit-Queue: Etienne Dechamps <edechamps@google.com>
Cr-Commit-Position: refs/heads/main@{#1418542}