Refactors the support of system AEC on Windows
This CL:

Removes usage of the Windows.Media.Effects.AudioEffectsManager UWP API
which was complex to use and only resulted in log messages that were not
utilized for any essential task.

Adds new methods if the EnforceSystemEchoCancellation flag is enabled.
These methods use IAudioEffectsManager::GetAudioEffects to enumerate all
available effects (instead of the old UWP API) and also checks that the
system AEC really could be enabled for the selected input device. If
not, a warning is logged.

Adds basic support of SetOutputDeviceForAec which now assists an active
AEC effect with the knowledge of that the default audio output device
should be used as reference stream for the AEC.

An example of a case where system AEC was enabled using a flag but the
selected input device did not support it is shown in
https://screenshot.googleplex.com/7ijVrnmRF8nWdpN Complete log:
https://paste.googleplex.com/4790780437463040

And sample where the device does support the AEC effect is shown in
https://screenshot.googleplex.com/98kb3NPSZC8VdkS Complete log:
https://paste.googleplex.com/4915036173369344

Bug: 393098901
Change-Id: Ic3361fdd0ef2eba3e934b85b3f795dcd30a79612
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6233979
Commit-Queue: Henrik Andreasson <henrika@chromium.org>
Reviewed-by: Olga Sharonova <olka@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1418560}