Reverse order of WebView saveState.
We want to be able to truncate WebView's saveState - omitting earlier
entries in the navigation history if it has grown too long. Given that
we don't know the size of a NavigationEntry before serializing it,
reversing the order of entries will allow us to truncate the list
without doing any extra copies.

Actually truncating the list will come in a follow up CL.

Bug: 389076708
Change-Id: Ic15def057d64284af7987dc91e0c5bea0e2e3d1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6252737
Reviewed-by: Peter Beverloo <peter@chromium.org>
Commit-Queue: Peter Conn <peconn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1418633}