[TabGroupSync] Add IPH for collaboration activity
The anchor can either be an individual tab or tab group header that
contains the local change. The instructions to lookup the anchor is
supplied by various callbacks in 2 separate listeners. This is to
cover when a tab is added, updated, moved, or removed from a shared
tab group.

The IPH uses the top container to lookup the browser window feature
which contains data about the most recent shared activity.

Bug: 370924453
Change-Id: I9b1df2ef2ac59b80847dd22e364bd009f388afe9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6207365
Commit-Queue: Mickey Burks <mickeyburks@chromium.org>
Reviewed-by: Darryl James <dljames@chromium.org>
Reviewed-by: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1418635}