user_agent_utils: Account for iOS in GetUserAgentInternal, support it in tests
After https://crrev.com/c/6242720, we no longer hardcode the UA string
based only on whether or not we are running on Android. This makes it
easier to support iOS as well.

Account for iOS in GetUserAgentInternal() for compatibility with
content's GetReducedUserAgent() and update the unit tests so that they
run correctly on iOS as well.

Bug: 389970327, 40252616
Change-Id: I10366d34d0fd8e7e3a18b818cf5896ff5eb5e0f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6246153
Reviewed-by: Mike Taylor <miketaylr@chromium.org>
Commit-Queue: Raphael Kubo da Costa <kubo@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1418647}