[BCIV] Proper fix for height animations
Just before an animation starts, the height is set to the final height, and all offsets in the renderer use the new height for reference. This also means that the OffsetTag values could be temporarily outside of the OffsetTagConstraints for some part of the animation.

A quick fix was implemented here https://chromium-review.googlesource.com/c/chromium/src/+/6149868

This cl removes the quick fix and fixes the problem properly by updating the OffsetTagConstraints properly to account for the full range of the animation, and also causes the browser to submit a new frame with the new height during the first frame of the animation.

Bug: 391867060
Change-Id: Ic6756fe213d16bad6d145e3b5ffd04bdba6c2efe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6192677
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Peilin Wang <peilinwang@google.com>
Reviewed-by: Bo Liu <boliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1418649}