Mac: add flag to skip sharing service for emailing links
This is a speculative fix for https://crbug.com/356643975.

Currently, email sharing is implemented via a macOS NSSharingService.
Sometimes, when a user invokes this with Command-Shift-I, an assert
fires deep in the bowels of ShareKit. Nobody on the team can repro
this at will, and we're not clear on if there's anything we're doing
to cause this.

This change adds a flag to do email shares directly, restoring (with
updates) the code that was removed in
4940014d71bd53b848c578f7a871e56ad6233120

NB: This introduces a new use of the banned `chrome::FindLastActive`
which is used extensively in share menu code. We can refactor this
for the whole class when this stabilizes, but I don't want to do
both at once.

No-Presubmit: true
Bug: 356643975
Change-Id: I5e6f28c9c347dc178dae15874c37825a79e628fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6248428
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Leonard Grey <lgrey@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1418674}