Disable errorprone on unused_resources target
The java code is copied from outside the chromium repo and CQ does not
enforce this at the moment because the errorprone bot is debug and this
is only built in release. Furthermore this could should be going away in
favour of R8 doing this for us. Disable the errorprone check for local
builds.

Change-Id: Ida281891a80b4eb130d168912a3fa41b4baa4c88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6252834
Auto-Submit: Mohamed Heikal <mheikal@chromium.org>
Commit-Queue: Sam Maier <smaier@chromium.org>
Commit-Queue: Mohamed Heikal <mheikal@chromium.org>
Reviewed-by: Sam Maier <smaier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1418689}