[PEPC] Relax font size restriction for the permission element
As detailed in the related bug, the permission element's font size check
is affected by the user's default font size, which is something that the
site cannot control. we check that the font has to be between the
equivalent values of ‘small’ and ‘xxx-large’, but these values actually
depend on the user’s chrome settings: chrome://settings/appearance

This means that what ‘small’ and ‘xxx-large’ actually get calculated to,
depends on the user’s settings. But font sizes provided in standard
units (e.g. 14px) do not. Because of that, font size settings of
“large”/”very large” actually make a 14px permission element invalid.
Since there are sites that don't set font sizes based on keywords but
rather have a fixed set of font sizes and let users use the browser zoom
function if they wish to enlarge the site, we want to make sure they are
supported as well.

A simple approach is to do the check twice, once as we do it today, and
once as if the user uses the “recommended” font size setting. If either
check succeeds then the font size is ok. This satisfies the condition
for sites that use a keyword (such as ‘large’) for the font size, as
well as sites that use absolute units (like px).

Pros: Simple solution design Easy for developers to understand (will
likely not need additional explanation) Cons: Expands the range of
allowed font sizes for users which use different than “medium” font
sizes (was unable to find UMA to elucidate which percentage of users
this is).

Also added some unit tests and removed an unused parameter.

Fixed: 374355283
Change-Id: I879d78247b118dd4fc995b5ac321680dc0fa6eed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6243801
Auto-Submit: Andy Paicu <andypaicu@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1418700}