Reformat some of the accessibility histograms file
This CL does not add any functional changes. We reformat part of the
accessibility histogram file to make it easier to read. The histogram
that tracks overall usage is split into 3, separating out the ChromeOS
specific histogram and the Windows zoom text histogram. In general the
ChromeOS histograms are in their own definitions, and zoom text does
not fit with the screenreader values of the original histogram, so
placing it in its own definition is a bit easier to parse/understand.
We add a bit more color to the descriptions and improve the formatting
so that the file is easier to quickly parse.

AX-Relnotes: N/A
Bug: 395095542
Change-Id: Iaecabd2080e874d8d919d5c86322ad406ac0d698
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6242825
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Auto-Submit: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1418713}