Remove Accessibility.WinSAToGo histogram
This CL removes the deprecated accessibility histogram that tracks the
usage of WinSAToGo. This histogram expired in 2020 but is still
recording data. The program has no usage, so we are not gaining any
useful data from this, and the histogram does not appear in the
histograms.xml file, so it does not appear by default in the histograms
dashboard.

Change-Id: I97ee3c4cf2651d61341e5d94e8e7b659765792da
AX-Relnotes: N/A
Bug: 395095542
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6244223
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Auto-Submit: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#1418717}