Remove the kProcessSharingWithStrictSiteInstances feature
In doing the SiteInstanceGroup work, we came across
kProcessSharingWithStrictSiteInstances, which has not been experimented
with for a long time, and can be removed. SiteInstanceGroups serves a
similar purpose in that it allows for multiple SiteInstances to be in
the same process.

This change also removes the existing SiteInstanceGroupManager, which
was a previous implementation of SiteInstanceGroup for the strict
SiteInstance mode. Usage of SiteInstanceGroupManager is gated behind
the strict SiteInstances flag, so it also hasn't been used or
experimented with for a long time, and can be removed.

Test: Removing a feature, remove strict SiteInstance checks in tests
Bug: 40612450, 40737467, 40161396, 390571607
Change-Id: I300523b8dabfe6de19dbf457eea143181e8e5587
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6050878
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Sharon Yang <yangsharon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1418735}