[search-engines] Update TemplateURLService::Update() to allow dual write
This CL enables dual-write to local and account values upon updates to
a search engine.

This adds a method UpdateData() to allow updating only the active data
(local or account). This is for the cases which should not
dual-write. For example, updating the last_visited time. OTOH, Update()
is now used for all the cases that require dual-write. For example,
if a search engine is edited by the user.

Bug: 374903497
Change-Id: I1ac884d1eb3049be0785e5b5a569f5f12f80c572
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6198269
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Auto-Submit: Ankush Singh <ankushkush@google.com>
Commit-Queue: Ankush Singh <ankushkush@google.com>
Cr-Commit-Position: refs/heads/main@{#1418769}