Implement LegacyMakeTypeface as its own IPC
LegacyMakeTypeface is used in a few places in the codebase, and one
important part of its contract that these places depend on is that it
should return the system default typeface if passed `nullptr` for the
`family_name` argument. In FontDataManager/FontDataServiceImpl, this
function is currently implemented as a call to `matchFamilyStyle`, which
doesn't respect this contract.

This CL implements LegacyMakeTypeface as a separate IPC on
`FontDataService`, which is the most likely approach to preserving
existing behavior. Arguably, `matchFamilyStyle` should also respect this
contract (see the comment for it in SkFontMgr), but since it's not the
case today we prefer decoupling this change from the act of making
FontDataManager functional.

Bug: None
Change-Id: I1350a2555f6e654eae3db59e95e922fcb31648ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6250221
Reviewed-by: David Yeung <dayeung@chromium.org>
Reviewed-by: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Anthony Vallée-Dubois <anthonyvd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1418803}