Reenable a flag to group child tab with parent tab in same group
This CL is trying to re-enable GroupNewTabWithParent. The previous
effort failed because crbug.com/394394717. The possible cause is
that the parent tab is in a group that is going away.
Will continue to keep an eye on this CL and see if it introduces
regression.

Bug: 379034423,394394717
Change-Id: I9c376050bfd07bef74ddd400a289eb0f5bb67fa9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6251716
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Min Qin <qinmin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1418825}