[customize chrome] Disable flaky categories test (instead of suite).
Test `error detection metrics fire correctly` is flaky (refer to bug
for previous runs). It fails because sometimes more metrics are fired
than it expects there to be. This CL makes it so this test is skipped,
and the rest of the tests are run.

Change-Id: I8376bc92fb8bd61d18022de556dee8da3e1352bf
Bug: 394712092
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6254177
Commit-Queue: Paul Adedeji <pauladedeji@google.com>
Reviewed-by: Marlon Facey <mfacey@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1418879}