[ios] Remove old MagicStack EG test
This removes an EG test that is no longer necessary because it tests the
non-compact SetUpList in the MagicStack and the behavior when all items
are completed. However, the SetUpList is now always show in "compact"
form. And there are separate tests for this.

Fixed: 41493926
Change-Id: I708adafb931fb444f3f2e7802ed0541dfcff15e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6248177
Reviewed-by: Benjamin Williams <bwwilliams@google.com>
Commit-Queue: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1418899}