Move `PrepareRequest` to `ContentAnalysisInfo`
This CL refactors `PrepareRequest` to be a member method of
`ContentAnalysisInfo` since most of the fields it sets (besides
analysis_connector) are known to that class. This will simplify the
implementation of other `RequestHandlerBase` sub-classes in future CLs.

Bug: 40839522
Change-Id: I036eb50b1a95ae33aded42a542e0e5ab6f4c78f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6253185
Commit-Queue: Nancy Xiao <nancylanxiao@google.com>
Reviewed-by: Nancy Xiao <nancylanxiao@google.com>
Auto-Submit: Dominique Fauteux-Chapleau <domfc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1418902}