Refactor build server to group build related work into a class
There is now a Build object/class that manages build specific work like
keeping track of build level statistics as well as creating and writing
to build logfiles.

Logfiles should no longer be leaked since we now can tell when a build
is "done" and close the logfile.

Title string is now based on global stats rather than per build to
better handle the case when two concurrent builds are updating the same
terminal title.

Added a global OptionsManager so as to not have to pipe the quiet option
everywhere.

Change-Id: I010982dc863ab45508957997c4ed8ad26902a4bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6249304
Commit-Queue: Mohamed Heikal <mheikal@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1418922}