Adding a class which will can retrieve Compute Engine metadata
This class will be used to call APIs from within the Compute Engine
Instance as the default service account which is associated with the
Instance. This will allow us to retrieve an Instance identity token,
generate an access token to call CRD APIs, and to check the scopes
for that access token for troubleshooting purposes.

Note that this initial version of the class only calls the HTTP
endpoint as that is guaranteed to exist on all Instances, the HTTPS
endpoint will require some additional work and is only available
on shielded Instances.

Bug: 388603116
Change-Id: I98346d1ad4ba089265529b3f3f29c3dbd57d4b44
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6250418
Reviewed-by: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Commit-Queue: Joe Downing <joedow@chromium.org>
Reviewed-by: Yuwei Huang <yuweih@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1418961}