[Gap Decorations]: Build Gap Intersection Points
This CL introduces the construction of gap intersection points for gaps
present in a grid. According to the spec[1], gap decorations are painted
relative to pairs of gap intersection points and are defined to exist
at: A) The center of an intersection between a gap and the content edge
of the container. B) The center of an intersection between gaps in
different directions.

Future changes will include the parsing of the `*rule-break` property
and the `*rule-outset` property. Following this, the painting code will
be updated to use the intersection pairs.

[1]: https://drafts.csswg.org/css-gaps-1/#layout-painting

Bug: 357648037
Change-Id: Ib8bc341a7a4d2f53dedbc18511205aa6d4ab984c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6175331
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Commit-Queue: Sam Davis Omekara <samomekarajr@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1418979}