android_deps: Configure 3pp to download the GMS TFLite packages
This change instructs the 3pp packager to create new CIPD packages for
the TensorFlow Lite in Google Play services library.

The CL depending on this may not land for a while (if ever) and so this
change will be reverted after the packages have been generated.

This is a reland of https://crrev.com/c/5141128 but with the dependency
updated to version 16.4.0 which is why the 3pp bot needs to be
triggered again.

Bug: 41486052
Change-Id: Ied437802f2d3dd988ed98a89df397729850c6d8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6254185
Auto-Submit: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1419026}