Add command line switch for opening glic on app startup
Use --glic-open-on-startup=attached to open glic in attached mode when
starting Chrome, or ...=detached (or any other value) to open detached.

This switch can be used for opening glic in automated tests until there
is a way to trigger the real entrypoints in a test environment.

Tested manually on Mac; no tests for this code path since it is itself
test support.

Bug: 391948342
Change-Id: Ifdc39cab5d47ec2e45a23e51e24db084038b6052
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6254881
Reviewed-by: Trevor Perrier <perrier@chromium.org>
Commit-Queue: Ian Wells <iwells@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1419044}